feat(substrate-bindings): Separate Enum from OutputEnum
