--- packages/substrate-bindings/src/types/GetEnum.ts
@@ -1,9 +1,0 @@
-import type { ExtractEnumValue, Enum } from "../codecs"
-
-export type GetEnum<T extends Enum<{ type: string; value: any }>> = {
-  [K in T["type"]]: (
-    ...args: ExtractEnumValue<T, K> extends undefined
-      ? []
-      : [value: ExtractEnumValue<T, K>]
-  ) => T
-}
